backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx
beautifulsoup4
requests
python-dotenv
python-pptx
rank_bm25
sqlalchemy
psycopg2-binary
chromadb
pydantic-settings
pydantic
